requirements.txt
torch>=1.7
numpy<1.21  # numba requires numpy<1.21,>=1.17
opencv-python
torchvision
scipy
tqdm
lmdb
pyyaml
tb-nightly
yapf
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0